Increase deassert delay for reset of Realtek phy
For the cases when the reset for the Realtek phy is issued, increase the
deassert delay to 100ms. In the power sequence app note, a minimum delay
of 72ms is stated after releasing the reset. Round up to 100ms to have a
certain buffer because the voltage ramp is not included in the 72ms
delay.