gpmi-nand-fus.c: Handle 0-bits in empty pages correctly
If an erased NAND page is not touched for a very long time and if there
are write accesses to neighboring pages, in rare cases also one or more
bits in the erased page may toggle from 1 to 0. For this reason, the
ECC engine of i.MX6 CPUs (BCH) can be told to also tolerate some
0-bits in empty (erased) pages. Then the page will still be detected
as empty despite the 0-bits.

Unfortunately, the ECC engine apparently does not correct these bits
back to 1, they stay at 0. This may cause filesystems like UBIFS to
fail with "corrupt empty space".

Add code to the NAND driver gpmi-nand-fus.c to handle these cases and
manually correct empty space to all 1-bits.