dts:F+S:imx8ulp: change clk-src for TPM4/5
Using SOSCDIV2 with a frequency of 24 MHz can cause problems with the
TPM module when changing frequencies.

This issue is not specific to a particular frequency, but rather to
changes between several frequencies, such as 1000 Hz <-> 2000 Hz.

Tests show that FROSC_DIV2 achieves more stable results at 48 MHz.

As TPM7 is intended for BL use, SOSCDIV2 remains here, as no frequency
changes are expected.