Update fsimx8mm M4 and remoteproc implementation
The way the Linux System comunicates with the M4-Core through remoteproc changed
somewhat over time. Because of this adjustments had to be made to the cm4 node
and its memory nodes in the devicetrees of the fsimx8mm platform. Add these
changes so that the M4 has exclusive access to the neccessary memory regions.