gpu:drm:bridge:tc358775: use framesync mode
Frame shifts were detected with the DSI interface of the iMX8ULP. By
default, the TC358775 uses the LineSync mode, which uses the HSS and VSS
signals of the DSI interface to trigger the Datastream for LVDS.

However, the LineSync mode shows a frame shift with each reset of the
DRM chain.

The FrameSync mode only triggers on VSS signals from the DSI interface
and uses HSFPR for timing, which significantly improves the frame during
DRM initialization. Only 4 shifted frames were detected after 500
resets.