Fix sgtl5000/sai clock racecondition
Sometimes the sgtl500 codec driver got probed before the sai driver
and claimed the SAI base clock. The the sai driver could not reparent
the lock anymore and error message.

To fix this, we move the clock assignment to the sound platform driver,
which should get loaded before codec and sai. This should make sure that
the clock is correctly configured, regardless which driver gets probed
first.