Fix fsimx8mp sgtl500 codec i2c issues
The SGTL5000 needs a extern clock to communicate over i2c.
This clock is provided by the SAI mclock on picocoremx8mp and efusmx8mp.
The SAI mclock in i.MX8MP needs to be additionally gated by the SAI
registers, which makes the load order of the drivers complicated.
Basically we need
SGTL enable clock -> SAI gate clock -> SGTL start i2c communication.

This is hard do map in the linux kernel driver structure, as SAI and
codec drivers are normally independent.

We introduce two new device tree properties:
1. SAI:fsl,sai-mclk-always-on to enable the mclk by default on probe
2. SGTL:  mclk-extern-enabled to defer the probe until the mclk is
enabled by the SAI driver.

This should male sure that the drivers are loaded in the correct order.